+++ app/src/main/java/com/example/aflami/designsystem/theme/Theme.kt
@@ -1,9 +1,9 @@
-package com.example.aflami.design_system.theme
+package com.example.aflami.designsystem.theme
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
-import com.example.aflami.design_system.textstyle.AflamiTextStyle
-import com.example.aflami.design_system.textstyle.LocalAflamiTextStyle
+import com.example.aflami.designsystem.textstyle.AflamiTextStyle
+import com.example.aflami.designsystem.textstyle.LocalAflamiTextStyle
 
 object Theme {
